fix useStorage
remove computed to useStorage file. The pinia state is reactive funciont and the computed propertie is not necesary.
Fix init http service
refactor authService with composables functions
the process property in vite it´s not neccesary, now use import.meta.
when use "defineProps" in component to instance in html this, its not necesario create a const porps = defineProps ,that only define when in script reuse the props variables.